+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/code/GraphNodePackage.java
@@ -242,4 +242,9 @@ public final class GraphNodePackage extends GraphNodeCode {
 	public String getCaption() {
 		return this.getPackageName();
 	}
+	
+	@Override
+	public String toString() {
+		return this.getCaption();
+	}
 }